requirements.txt
dash
plotly
pandas
scikit-learn
numpy
scipy
numba
umap-learn
dash-bootstrap-components
pyDOE
matplotlib
kaleido